+++ packages/lib/utils/vue3Utils.js
@@ -3,13 +3,11 @@
  */
 
 import { defineComponent, h, resolveComponent as _resolveComponent } from 'vue';
+import { pathSeparator } from './vueCommonUtils';
 
 export {
     nodePath2ClassName, isRootNodePath, computedCurPath, getPathVal, path2prop
-} from './vueUtils';
-
-// 内部使用 . ，配置数据key不能出现.
-const pathSeparator = '.';
+} from './vueCommonUtils';
 
 // 删除当前path值
 export function deletePathVal(vueData, name) {
